requirements.txt
livekit==0.19.1
livekit-api==0.8.1
livekit-agents==0.12.10
livekit-plugins-deepgram==0.6.16
livekit-plugins-openai==0.10.16
livekit-plugins-silero==0.7.4
livekit-plugins-turn-detector==0.3.6
livekit-plugins-rag
fastapi==0.104.0
pydantic>=2.0.2,<3.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
annoy>=1.17.0
numpy>=1.26.0,<2.0.0
torch>=2.1.0
agentops
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.7.0
isort==5.12.0
mypy==1.5.0
beautifulsoup4
aiofiles
